feat: дописал тесты

+++ src/FileService.Contracts/StartMultipartUploadResponse.cs
@@ -4,4 +4,4 @@ public record StartMultipartUploadResponse(
     Guid MediaAssetId,
     string UploadId, 
     IReadOnlyList<ChunkUploadUrl> ChunkUploadUrls,
-    long ChunkSize);
+    int ChunkSize);
